requirements.txt
streamlit
openai
pyautogen
fpdf
